// src/app/core/interfaces/helpers/aspx-inputs/admin/c-pass-form-aspx-input.ts
export interface CPassFormAspxInput {
  secretQuestion: HTMLSelectElement;
  answer: HTMLInputElement;
  newPassword: HTMLInputElement;
  confirmPassword: HTMLInputElement;
  capcha: HTMLInputElement;
  capchaImage: HTMLImageElement;
  capchaButton: HTMLInputElement;
  submitButton: HTMLInputElement;
}